requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
nbconvert>=7.0.0
nbformat>=5.7.0

# Markdown processing
markdown-it-py>=3.0.0
myst-parser>=2.0.0

# Image handling
pillow>=10.0.0

# Accessibility (optional)
# pa11y requires Node.js/npm installation
